requirements.txt
keyboard
mouse
PyYAML